# swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/users":{"post":{"operationId":"UsersController_create","summary":"Admin create new user","description":"\n* Only admin can use this API\n\n* Admin create user and give some specific information","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["users"]}},"/api/user-roles":{"post":{"operationId":"UserRolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRoleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRole"}}}}},"tags":["user-roles"]},"get":{"operationId":"UserRolesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["user-roles"]}},"/api/user-roles/{id}":{"get":{"operationId":"UserRolesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRole"}}}}},"tags":["user-roles"]},"patch":{"operationId":"UserRolesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRole"}}}}},"tags":["user-roles"]},"delete":{"operationId":"UserRolesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["user-roles"]}},"/api/collections":{"post":{"operationId":"CollectionsController_create","summary":"User create their collection","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"name":{"type":"string","default":"Learn Kitchen Vocabulary"},"description":{"type":"string","default":"Some description"},"level":{"type":"string","enum":["easy","medium","hard","chaos"],"default":"chaos"},"is_public":{"type":"boolean","default":true},"image":{"type":"string","format":"binary"}},"required":["name","level","is_public","image"]}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}}},"tags":["collections"],"security":[{"token":[]}]},"get":{"operationId":"CollectionsController_findAll","parameters":[{"name":"offset","required":true,"in":"query","examples":{"0":{"value":0,"description":"Start from 0"},"10":{"value":10,"description":"Skip 10 collection"}},"schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","examples":{"10":{"value":10,"description":"Get 10 collection"},"50":{"value":50,"description":"Get 50 collection"}},"schema":{"type":"number"}},{"name":"level","required":false,"in":"query","examples":{"one_level_type":{"value":["hard"]},"two_level_type":{"value":["easy","medium"]}},"schema":{"type":"array"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["collections"]}},"/api/collections/{id}":{"get":{"operationId":"CollectionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}}},"tags":["collections"]},"patch":{"operationId":"CollectionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCollectionDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}}},"tags":["collections"]},"delete":{"operationId":"CollectionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["collections"]}},"/api/topics":{"post":{"operationId":"TopicsController_create","summary":"Admin create topic","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"name":{"type":"string","default":"Learn Kitchen Vocabulary"},"description":{"type":"string","default":"Some description"},"images":{"type":"array","items":{"type":"string","format":"binary"}}},"required":["name","images"]}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Topic"}}}}},"tags":["topics"]},"get":{"operationId":"TopicsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["topics"]}},"/api/topics/{id}":{"get":{"operationId":"TopicsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Topic"}}}}},"tags":["topics"]},"patch":{"operationId":"TopicsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTopicDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Topic"}}}}},"tags":["topics"]},"delete":{"operationId":"TopicsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["topics"]}},"/api/flash-cards":{"post":{"operationId":"FlashCardsController_create","summary":"User create their new flash card","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"vocabulary":{"type":"string","default":"provision"},"definition":{"type":"string","default":"the action of providing or supplying something for use."},"meaning":{"type":"string","default":"sự cung cấp"},"pronunciation":{"type":"string","default":"prəˈviZHən"},"examples[]":{"type":"array","items":{"type":"string","default":""},"default":["new contracts for the provision of services","low levels of social provision","civilian contractors were responsible for provisioning these armies"]},"image":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlashCard"}}}}},"tags":["flash-cards"],"security":[{"token":[]}]},"get":{"operationId":"FlashCardsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["flash-cards"]}},"/api/flash-cards/{id}":{"get":{"operationId":"FlashCardsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlashCard"}}}}},"tags":["flash-cards"]},"patch":{"operationId":"FlashCardsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFlashCardDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlashCard"}}}}},"tags":["flash-cards"]},"delete":{"operationId":"FlashCardsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["flash-cards"]}},"/api/auth/sign-up":{"post":{"operationId":"AuthController_signUp","summary":"User sign up to platform","description":"## User sign up","servers":[{"url":"http://localhost:8080","description":"Current server"},{"url":"http://localhost:9000","description":"Authentication service if exist"}],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"},"examples":{"user_1":{"value":{"first_name":"John","last_name":"Doe","email":"johndoe@example.com","password":"User@123"}},"user_2":{"value":{"first_name":"Michael","last_name":"Smith","email":"michaelsmith@example.com","password":"User@123"}}}}}},"responses":{"201":{"description":"User created successfully!!","content":{"application/json":{"examples":{"created_user":{"summary":"Response after sign up","value":{"access_token":"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjQ0MWNkNmJlMWQ0ZTBiNDRjNzA3NDk2IiwiaWF0IjoxNjgyMDM0MDI3LCJleHAiOjE2ODIwMzc2Mjd9.AH4z7uDWuEDjOs8sesB0ItxKUJ2M3rjul1D1mmjAKieOZblej5mp0JQE5IdgB9LlAOzOtKOLL5RWhxLCZ-YskvoRA7Yqza_rOjfIHeNseC3M66kKYqORN07aZDiA2OWhT3pXBqoKRCUBQCKLgMCAPT-CHryc0wUQGaKxP8YJO8dwIhGtjADchmzNJVBs4G7qYnpZAsORayd5GNfgoLpWmVFIBHSnPLNIL4dL8dLof0GBmVhdjhnHIUXYQlqL1wiwsmxmUC9TU2uiChm-TAhuiQyVwFokSySBJzBrLmEtgy89aaR0YizFK-QMg2xW3cJiiRzEBigTdsR0kvdUlk5GOg","refresh_token":"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjQ0MWNkNmJlMWQ0ZTBiNDRjNzA3NDk2IiwiaWF0IjoxNjgyMDM0MDI3LCJleHAiOjE2ODIwNTkyMjd9.aKNZymKdf3VEbPkda2cYYTS7KlpCbTqdXP30LREQ2b_fJ8q8cA0OyNEARK3Jm5yGsKoNd3txi54XmEbf19LC9CuDf9kwgLasPizEeMZsAJqSbSguzE4-9b4sSdf22GyipCcZJpkXkp01Bew04J8Y4FqhNARONsWzySXg8_VVWOGkfHGJVHFs7xYyVvmt3RErJwRM5s1Ou1ok7VW62FSTSAvXw6-qsHp5T7kXo73jECBqSuNEs5JcdluoBjdaAxggHYaDgTXoRh7y4Mn_fVKCQarAsUAxg6w0fxc8Gj0nP1ct3-GjG-Of-0O-iF7uynI2Lnq_On7WUsH7rFSysNyHUg"}}}}}},"400":{"description":"Validation failed","content":{"application/json":{"examples":{"invalid_email_password":{"value":{"statusCode":400,"message":["email must be an email","password is not strong enough"],"error":"Bad Request"}},"some_fields_missing":{"value":{"statusCode":400,"message":["last_name must be shorter than or equal to 50 characters","last_name should not be empty"],"error":"Bad Request"}}}}}},"409":{"description":"Conflict user info","content":{"application/json":{"examples":{"email_duplication":{"value":{"statusCode":409,"message":"Email already existed!!","error":"Conflict"}},"username_duplication":{"value":{"statusCode":409,"message":"Username already existed!!","error":"Conflict"}}}}}}},"tags":["auth"]}},"/api/auth/sign-in":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"},"examples":{"user_1":{"value":{"first_name":"John","last_name":"Doe","email":"johndoe@example.com","password":"User@123"}},"user_2":{"value":{"first_name":"Michael","last_name":"Smith","email":"michaelsmith@example.com","password":"User@123"}}}}}},"responses":{"201":{"description":""},"401":{"description":"Unauthorized","content":{"application/json":{"example":{"statusCode":400,"message":"Wrong credentials!!","error":"Bad Request"}}}}},"tags":["auth"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refreshAccessToken","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}}},"info":{"title":"Flash card project","description":"## The flash card API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateAddressDto":{"type":"object","properties":{"street":{"type":"string","minLength":2,"maxLength":120},"state":{"type":"string","minLength":2,"maxLength":50},"city":{"type":"string","minLength":2,"maxLength":50},"postal_code":{"type":"number"},"country":{"type":"string","minLength":2,"maxLength":50}},"required":["state","city","postal_code","country"]},"CreateUserDto":{"type":"object","properties":{"first_name":{"type":"string","maxLength":50},"last_name":{"type":"string","maxLength":50},"email":{"type":"string","maxLength":50},"username":{"type":"string","maxLength":50},"phone_number":{"type":"string"},"password":{"type":"string"},"address":{"type":"array","items":{"$ref":"#/components/schemas/CreateAddressDto"}},"interested_languages":{"type":"array","items":{"type":"string","enum":["English","French","Japanese","Korean","Spanish"]}}},"required":["first_name","last_name","email","username","password"]},"UserRole":{"type":"object","properties":{"name":{"type":"string"},"_description":{"type":"string"},"id":{"type":"string"},"deleted_at":{"format":"date-time","type":"string"}},"required":["name","_description","deleted_at"]},"Address":{"type":"object","properties":{"street":{"type":"string"},"state":{"type":"string"},"city":{"type":"string"},"postal_code":{"type":"string"},"country":{"type":"string"},"id":{"type":"string"},"deleted_at":{"format":"date-time","type":"string"}},"required":["state","city","country","deleted_at"]},"DailyCheckIn":{"type":"object","properties":{"checked_date":{"format":"date-time","type":"string"},"access_amount":{"type":"number"},"eligible_for_reward":{"type":"boolean"},"reward_days_count":{"type":"number"}},"required":["checked_date","access_amount","eligible_for_reward","reward_days_count"]},"User":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"phone_number":{"type":"string"},"password":{"type":"string"},"avatar":{"type":"string"},"date_of_birth":{"format":"date-time","type":"string"},"gender":{"type":"string"},"point":{"type":"number"},"role":{"$ref":"#/components/schemas/UserRole"},"headline":{"type":"string"},"friendly_id":{"type":"number"},"address":{"type":"array","items":{"$ref":"#/components/schemas/Address"}},"default_address":{"type":"string"},"stripe_customer_id":{"type":"string"},"current_refresh_token":{"type":"string"},"daily_check_in":{"type":"array","items":{"$ref":"#/components/schemas/DailyCheckIn"}},"last_check_in":{"format":"date-time","type":"string"},"last_get_check_in_rewards":{"format":"date-time","type":"string"},"id":{"type":"string"},"deleted_at":{"format":"date-time","type":"string"}},"required":["first_name","last_name","email","phone_number","password","avatar","date_of_birth","gender","point","role","headline","friendly_id","address","stripe_customer_id","current_refresh_token","last_check_in","last_get_check_in_rewards","deleted_at"]},"UpdateUserDto":{"type":"object","properties":{"phone_number":{"type":"string"},"date_of_birth":{"format":"date-time","type":"string"},"gender":{"type":"string","enum":["MALE","FEMALE","OTHER"]},"headline":{"type":"string","maxLength":200},"first_name":{"type":"string","maxLength":50},"last_name":{"type":"string","maxLength":50},"address":{"type":"array","items":{"$ref":"#/components/schemas/CreateAddressDto"}},"interested_languages":{"type":"array","items":{"type":"string","enum":["English","French","Japanese","Korean","Spanish"]}}}},"CreateUserRoleDto":{"type":"object","properties":{}},"UpdateUserRoleDto":{"type":"object","properties":{}},"Collection":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"level":{"enum":["easy","medium","hard","chaos"],"type":"string"},"order":{"type":"number"},"image":{"type":"string"},"total_flash_cards":{"type":"number"},"is_public":{"type":"boolean"},"user":{"$ref":"#/components/schemas/User"},"id":{"type":"string"},"deleted_at":{"format":"date-time","type":"string"}},"required":["name","description","level","order","image","total_flash_cards","is_public","user","deleted_at"]},"UpdateCollectionDto":{"type":"object","properties":{}},"Topic":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"id":{"type":"string"},"deleted_at":{"format":"date-time","type":"string"}},"required":["name","description","deleted_at"]},"UpdateTopicDto":{"type":"object","properties":{}},"FlashCard":{"type":"object","properties":{"vocabulary":{"type":"string"},"image":{"type":"string"},"definition":{"type":"string"},"meaning":{"type":"string"},"pronunciation":{"type":"string"},"examples":{"type":"array","items":{"type":"string"}},"is_public":{"type":"boolean"},"user":{"$ref":"#/components/schemas/User"},"topics":{"type":"array","items":{"$ref":"#/components/schemas/Topic"}},"id":{"type":"string"},"deleted_at":{"format":"date-time","type":"string"}},"required":["vocabulary","image","definition","meaning","examples","is_public","user","topics","deleted_at"]},"UpdateFlashCardDto":{"type":"object","properties":{}},"SignUpDto":{"type":"object","properties":{"first_name":{"type":"string","maxLength":50},"last_name":{"type":"string","maxLength":50},"email":{"type":"string"},"password":{"type":"string"}},"required":["first_name","last_name","email","password"]}}}}